[ecore][ts] add AbstractEObject class and define attributes in BasicEObject
