+++ ecore/src/BasicEObject.ts
@@ -8,19 +8,11 @@
 // *****************************************************************************
 
 import {
-    EcoreUtils,
-    ETreeIterator,
     ENotification,
-    ENotificationChain,
     EObject,
     EObjectList,
-    ENotifyingList,
-    EObjectInternal,
     EList,
     EClass,
-    EOperation,
-    EAttribute,
-    EReference,
     EResource,
     EStructuralFeature,
     ImmutableEList,
@@ -29,19 +21,10 @@ import {
     BasicNotifier,
     EventType,
     EOPPOSITE_FEATURE_BASE,
+    AbstractEObject,
+    EAdapter,
+    AbstractENotifierList,
 } from "./internal";
-
-export function isEReference(s: EStructuralFeature): s is EReference {
-    return "eReferenceType" in s;
-}
-
-export function isEAttribute(s: EStructuralFeature): s is EAttribute {
-    return "eAttributeType" in s;
-}
-
-function isNumeric(n) {
-    return !isNaN(parseFloat(n)) && isFinite(n);
-}
 
 type getFeatureFnType = (c: EClass) => EList<EStructuralFeature>;
 
@@ -161,72 +144,41 @@ class ContentsListAdapter extends AbstractEAdapter {
     }
 }
 
-export class BasicEObject extends BasicNotifier implements EObjectInternal {
-    private _eResource: EResource;
-    private _eContainer: EObject;
+export class BasicEObject extends AbstractEObject {
+    private _eResource: EResource = null;
+    private _eContainer: EObject = null;
     private _eContainerFeatureID: number = -1;
-    private _eProxyURI: URL;
-    private _contentsListAdapter: ContentsListAdapter;
-    private _crossReferencesListAdapter: ContentsListAdapter;
+    private _eProxyURI: URL = null;
+    private _contentsListAdapter: ContentsListAdapter = null;
+    private _crossReferencesListAdapter: ContentsListAdapter = null;
+    private _adapters: EList<EAdapter> = null;
+    private _deliver: boolean = true;
 
-    constructor() {
-        super();
+    get eAdapters(): EList<EAdapter> {
+        if (!this._adapters) {
+            this._adapters = new AbstractENotifierList(this);
+        }
+        return this._adapters;
     }
 
-    eClass(): EClass {
-        return this.eStaticClass();
+    get eDeliver(): boolean {
+        return this._deliver;
     }
 
-    eStaticClass(): EClass {
-        return null;
+    set eDeliver(deliver: boolean) {
+        this._deliver = deliver;
+    }
+
+    eBasicAdapters(): EList<EAdapter> {
+        return this._adapters;
     }
 
     eInternalContainer(): EObject {
         return this._eContainer;
     }
 
-    eContainer(): EObject {
-        let eContainer = this._eContainer;
-        if (eContainer && eContainer.eIsProxy()) {
-            let resolved = this.eResolveProxy(eContainer);
-            if (resolved != eContainer) {
-                let notifications = this.eBasicRemoveFromContainer(null);
-                this._eContainer = resolved;
-                if (notifications) {
-                    notifications.dispatch();
-                }
-                if (
-                    this.eNotificationRequired &&
-                    this._eContainerFeatureID >= EOPPOSITE_FEATURE_BASE
-                ) {
-                    this.eNotify(
-                        new Notification(
-                            this,
-                            EventType.RESOLVE,
-                            this._eContainerFeatureID,
-                            eContainer,
-                            resolved
-                        )
-                    );
-                }
-            }
-            return resolved;
-        }
-        return eContainer;
-    }
-
-    eContainerFeatureID(): number {
+    eInternalContainerFeatureID(): number {
         return this._eContainerFeatureID;
-    }
-
-    eResource(): EResource {
-        let resource = this._eResource;
-        if (!resource) {
-            if (this._eContainer) {
-                resource = this._eContainer.eResource();
-            }
-        }
-        return resource;
     }
 
     eInternalResource(): EResource {
@@ -237,79 +189,9 @@ export class BasicEObject extends BasicNotifier implements EObjectInternal {
         this._eResource = eResource;
     }
 
-    eSetResource(newResource: EResource, n: ENotificationChain): ENotificationChain {
-        let notifications = n;
-        let oldResource = this.eInternalResource();
-        if (oldResource && newResource) {
-            let list = oldResource.eContents() as ENotifyingList<EObject>;
-            notifications = list.removeWithNotification(this, notifications);
-            oldResource.detached(this);
-        }
-        if (this._eContainer) {
-            if (this.eContainmentFeature().isResolveProxies) {
-                let oldContainerResource = this._eContainer.eResource();
-                if (oldContainerResource) {
-                    if (!newResource) {
-                        oldContainerResource.attached(this);
-                    } else if (!oldResource) {
-                        oldContainerResource.detached(this);
-                    }
-                }
-            } else {
-                notifications = this.eBasicRemoveFromContainer(notifications);
-                notifications = this.eBasicSetContainer(null, -1, notifications);
-            }
-        }
-        this.eSetInternalResource(newResource);
-        return notifications;
-    }
-    eContainingFeature(): EStructuralFeature {
-        if (this._eContainer) {
-            if (this._eContainerFeatureID <= EOPPOSITE_FEATURE_BASE) {
-                let feature = <EStructuralFeature>(
-                    this._eContainer
-                        .eClass()
-                        .getEStructuralFeature(EOPPOSITE_FEATURE_BASE - this._eContainerFeatureID)
-                );
-                return feature;
-            } else {
-                let reference = <EReference>(
-                    this.eClass().getEStructuralFeature(this._eContainerFeatureID)
-                );
-                return reference.eOpposite;
-            }
-        }
-        return null;
-    }
-
-    eContainmentFeature(): EReference {
-        return this.eObjectContainmentFeature(this, this._eContainer, this._eContainerFeatureID);
-    }
-
-    private eObjectContainmentFeature(
-        o: EObject,
-        container: EObject,
-        containerFeatureID: number
-    ): EReference {
-        if (this._eContainer) {
-            if (this._eContainerFeatureID <= EOPPOSITE_FEATURE_BASE) {
-                let feature: EStructuralFeature = this._eContainer
-                    .eClass()
-                    .getEStructuralFeature(EOPPOSITE_FEATURE_BASE - containerFeatureID);
-                if (isEReference(feature)) {
-                    return feature;
-                }
-            } else {
-                let feature: EStructuralFeature = this.eClass().getEStructuralFeature(
-                    containerFeatureID
-                );
-                if (isEReference(feature)) {
-                    return feature;
-                }
-            }
-            throw new Error("The containment feature could not be located");
-        }
-        return null;
+    eSetInternalContainer(container: EObject, containerFeatureID: number): void {
+        this._eContainer = container;
+        this._eContainerFeatureID = containerFeatureID;
     }
 
     eContents(): EList<EObject> {
@@ -323,16 +205,6 @@ export class BasicEObject extends BasicNotifier implements EObjectInternal {
         return this._contentsListAdapter.getList();
     }
 
-    eAllContents(): IterableIterator<EObject> {
-        return new ETreeIterator<EObject, EObject>(
-            this,
-            false,
-            function (o: EObject): Iterator<EObject> {
-                return o.eContents()[Symbol.iterator]();
-            }
-        );
-    }
-
     eCrossReferences(): EList<EObject> {
         if (!this._crossReferencesListAdapter)
             this._crossReferencesListAdapter = new ContentsListAdapter(
@@ -342,302 +214,6 @@ export class BasicEObject extends BasicNotifier implements EObjectInternal {
                 }
             );
         return this._crossReferencesListAdapter.getList();
-    }
-
-    eFeatureID(feature: EStructuralFeature): number {
-        if (!this.eClass().eAllStructuralFeatures.contains(feature))
-            throw new Error("The feature '" + feature.name + "' is not a valid feature");
-        return this.eDerivedFeatureID(feature.eContainer(), feature.featureID);
-    }
-
-    eDerivedFeatureID(container: EObject, featureID: number): number {
-        return featureID;
-    }
-
-    eOperationID(operation: EOperation): number {
-        if (!this.eClass().eAllOperations.contains(operation)) {
-            throw new Error("The operation '" + operation.name + "' is not a valid feature");
-        }
-
-        return this.eDerivedFeatureID(operation.eContainer(), operation.operationID);
-    }
-
-    eDerivedOperationID(container: EObject, operationID: number): number {
-        return operationID;
-    }
-
-    eGet(feature: EStructuralFeature): any {
-        return this.eGetFromFeature(feature, true);
-    }
-
-    eGetResolve(feature: EStructuralFeature, resolve: boolean): any {
-        return this.eGetFromFeature(feature, resolve);
-    }
-
-    private eGetFromFeature(feature: EStructuralFeature, resolve: boolean): any {
-        let featureID = this.eFeatureID(feature);
-        if (featureID >= 0) {
-            return this.eGetFromID(featureID, resolve);
-        }
-        throw new Error("The feature '" + feature.name + "' is not a valid feature");
-    }
-
-    eGetFromID(featureID: number, resolve: boolean): any {
-        let feature = this.eClass().getEStructuralFeature(featureID);
-        if (!feature) {
-            throw new Error("Invalid featureID: " + featureID);
-        }
-        return null;
-    }
-
-    eSet(feature: EStructuralFeature, newValue: any): void {
-        let featureID = this.eFeatureID(feature);
-        if (featureID >= 0) {
-            this.eSetFromID(featureID, newValue);
-        } else {
-            throw new Error("The feature '" + feature.name + "' is not a valid feature");
-        }
-    }
-
-    eSetFromID(featureID: number, newValue: any): void {
-        let feature = this.eClass().getEStructuralFeature(featureID);
-        if (!feature) {
-            throw new Error("Invalid featureID: " + featureID);
-        }
-    }
-
-    eIsSet(feature: EStructuralFeature): boolean {
-        let featureID = this.eFeatureID(feature);
-        if (featureID >= 0) {
-            return this.eIsSetFromID(featureID);
-        }
-        throw new Error("The feature '" + feature.name + "' is not a valid feature");
-    }
-
-    eIsSetFromID(featureID: number): boolean {
-        let feature = this.eClass().getEStructuralFeature(featureID);
-        if (!feature) {
-            throw new Error("Invalid featureID: " + featureID);
-        }
-        return false;
-    }
-
-    eUnset(feature: EStructuralFeature): void {
-        let featureID = this.eFeatureID(feature);
-        if (featureID >= 0) {
-            this.eUnsetFromID(featureID);
-        } else {
-            throw new Error("The feature '" + feature.name + "' is not a valid feature");
-        }
-    }
-
-    eUnsetFromID(featureID: number): void {
-        let feature = this.eClass().getEStructuralFeature(featureID);
-        if (!feature) {
-            throw new Error("Invalid featureID: " + featureID);
-        }
-    }
-
-    eInvoke(operation: EOperation, args: EList<any>): any {
-        let operationID = this.eOperationID(operation);
-        if (operationID >= 0) {
-            return this.eInvokeFromID(operationID, args);
-        }
-        throw new Error("The operation '" + operation.name + "' is not a valid operation");
-    }
-
-    eInvokeFromID(operationID: number, args: EList<any>): any {
-        let operation = this.eClass().getEOperation(operationID);
-        if (!operation) {
-            throw new Error("Invalid operationID: " + operationID);
-        }
-    }
-
-    eInverseAdd(otherEnd: EObject, featureID: number, n: ENotificationChain): ENotificationChain {
-        let notifications = n;
-        if (featureID >= 0) {
-            this.eBasicInverseAdd(otherEnd, featureID, notifications);
-        } else {
-            notifications = this.eBasicRemoveFromContainer(notifications);
-            return this.eBasicSetContainer(otherEnd, featureID, notifications);
-        }
-    }
-
-    eBasicInverseAdd(
-        otherEnd: EObject,
-        featureID: number,
-        notifications: ENotificationChain
-    ): ENotificationChain {
-        return notifications;
-    }
-
-    eInverseRemove(
-        otherEnd: EObject,
-        featureID: number,
-        notifications: ENotificationChain
-    ): ENotificationChain {
-        return featureID >= 0
-            ? this.eBasicInverseRemove(otherEnd, featureID, notifications)
-            : this.eBasicSetContainer(null, featureID, notifications);
-    }
-
-    eBasicInverseRemove(
-        otherEnd: EObject,
-        featureID: number,
-        notifications: ENotificationChain
-    ): ENotificationChain {
-        return notifications;
-    }
-
-    protected eBasicSetContainer(
-        newContainer: EObject,
-        newContainerFeatureID: number,
-        n: ENotificationChain
-    ): ENotificationChain {
-        let notifications = n;
-        let oldResource = this._eResource;
-        let oldContainer = this._eContainer;
-        let oldContainerFeatureID = this._eContainerFeatureID;
-
-        let newResource: EResource = null;
-        if (oldResource) {
-            if (
-                newContainer &&
-                !this.eObjectContainmentFeature(this, newContainer, newContainerFeatureID)
-                    .isResolveProxies
-            ) {
-                let list = oldResource.eContents() as ENotifyingList<EObject>;
-                notifications = list.removeWithNotification(this, notifications);
-                this.eSetInternalResource(null);
-                newResource = newContainer.eResource();
-            } else {
-                oldResource = null;
-            }
-        } else {
-            if (oldContainer) {
-                oldResource = oldContainer.eResource();
-            }
-            if (newContainer) {
-                newResource = newContainer.eResource();
-            }
-        }
-
-        if (oldResource && oldResource != newResource) {
-            oldResource.detached(this);
-        }
-
-        if (newResource && newResource && oldResource) {
-            newResource.attached(this);
-        }
-
-        // basic set
-        this._eContainer = newContainer;
-        this._eContainerFeatureID = newContainerFeatureID;
-
-        // notification
-        if (this.eNotificationRequired) {
-            if (
-                oldContainer != null &&
-                oldContainerFeatureID >= 0 &&
-                oldContainerFeatureID != newContainerFeatureID
-            ) {
-                let notification = new Notification(
-                    this,
-                    EventType.SET,
-                    oldContainerFeatureID,
-                    oldContainer,
-                    null
-                );
-                if (notifications != null) {
-                    notifications.add(notification);
-                } else {
-                    notifications = notification;
-                }
-            }
-            if (newContainerFeatureID >= 0) {
-                let notification = new Notification(
-                    this,
-                    EventType.SET,
-                    newContainerFeatureID,
-                    oldContainerFeatureID == newContainerFeatureID ? oldContainer : null,
-                    newContainer
-                );
-                if (notifications != null) {
-                    notifications.add(notification);
-                } else {
-                    notifications = notification;
-                }
-            }
-        }
-        return notifications;
-    }
-
-    eBasicRemoveFromContainer(notifications: ENotificationChain): ENotificationChain {
-        if (this._eContainerFeatureID >= 0)
-            return this.eBasicRemoveFromContainerFeature(notifications);
-        else {
-            if (this._eContainer != null)
-                return this.eInverseRemove(
-                    this,
-                    EOPPOSITE_FEATURE_BASE - this._eContainerFeatureID,
-                    notifications
-                );
-        }
-        return notifications;
-    }
-
-    eBasicRemoveFromContainerFeature(notifications: ENotificationChain): ENotificationChain {
-        let feature = this.eClass().getEStructuralFeature(this._eContainerFeatureID);
-        if (isEReference(feature)) {
-            let inverseFeature = feature.eOpposite;
-            if (this._eContainer != null && inverseFeature != null)
-                return (this._eContainer as EObjectInternal).eInverseRemove(
-                    this,
-                    inverseFeature.featureID,
-                    notifications
-                );
-        }
-        return notifications;
-    }
-
-    eObjectForFragmentSegment(fragment: string): EObject {
-        let index = -1;
-        if (fragment && fragment.length > 0 && isNumeric(fragment.charAt(fragment.length - 1))) {
-            index = fragment.lastIndexOf(".");
-            if (index != -1) {
-                let pos = parseInt(fragment.slice(index + 1));
-                let eFeatureName = fragment.slice(1, index);
-                let eFeature = this.getStructuralFeatureFromName(eFeatureName);
-                let list = this.eGetResolve(eFeature, false) as EList<EObject>;
-                if (pos < list.size()) {
-                    return list.get(pos);
-                }
-            }
-        }
-        if (index == -1) {
-            let eFeature = this.getStructuralFeatureFromName(fragment);
-            return this.eGetResolve(eFeature, false) as EObject;
-        }
-        return null;
-    }
-
-    eURIFragmentSegment(feature: EStructuralFeature, o: EObject): string {
-        let s = "@";
-        s += feature.name;
-        if (feature.isMany) {
-            let v = this.eGetResolve(feature, false);
-            let i = (v as EList<EObject>).indexOf(o);
-            s += "." + i.toString();
-        }
-        return s;
-    }
-
-    private getStructuralFeatureFromName(featureName: string): EStructuralFeature {
-        let eFeature = this.eClass().getEStructuralFeatureFromName(featureName);
-        if (!eFeature) {
-            throw new Error("The feature " + featureName + " is not a valid feature");
-        }
-        return eFeature;
     }
 
     eIsProxy(): boolean {
@@ -650,9 +226,5 @@ export class BasicEObject extends BasicNotifier implements EObjectInternal {
 
     eSetProxyURI(uri: URL): void {
         this._eProxyURI = uri;
-    }
-
-    eResolveProxy(proxy: EObject): EObject {
-        return EcoreUtils.resolveInObject(proxy, this);
     }
 }
