+++ ecore/src/EClassExt.ts
@@ -21,7 +21,12 @@ import {
     AbstractEAdapter,
     EventType,
     EcoreConstants,
+    EClassifier,
 } from "./internal";
+
+export function isEClass(e: EClassifier): e is EClass {
+    return "isAbstract" in e;
+}
 
 class ESuperAdapter extends AbstractEAdapter {
     constructor(private _eClass: EClassExt) {
