+++ ecore/src/EDataTypeExt.ts
@@ -1,0 +1,5 @@
+import { EClassifier, EDataType } from "./internal";
+
+export function isEDataType(e: EClassifier): e is EDataType {
+    return "isSerializable" in e;
+}
