+++ ecore/src/internal.ts
@@ -55,13 +55,14 @@ export * from "./EcoreConstants";
 export * from "./EcorePackage";
 export * from "./EcoreFactory";
 
-export * from "./AbstractEAdapter";
 export * from "./AbstractEList";
-export * from "./AbstractENotifier";
+export * from "./BasicEList";
+export * from "./AbstractEAdapter";
 export * from "./AbstractNotification";
+export * from "./AbstractENotifier";
+export * from "./AbstractEObject";
 export * from "./Notification";
 export * from "./NotificationChain";
-export * from "./BasicEList";
 export * from "./ImmutableEList";
 export * from "./AbstractNotifyingList";
 export * from "./BasicNotifier";
@@ -86,6 +87,7 @@ export * from "./EClassifierExt";
 export * from "./EClassImpl";
 export * from "./EClassExt";
 export * from "./EDataTypeImpl";
+export * from "./EDataTypeExt";
 export * from "./EEnumImpl";
 export * from "./EEnumLiteralImpl";
 export * from "./EFactoryImpl";
